requirements.txt
absl-py==1.3.0
appnope==0.1.3
apriltag==0.0.16
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dlib==19.24.0
dnspython==2.2.1
executing==1.2.0
filelock==3.8.2
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
gdown==4.6.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.2.0
ipykernel==6.20.2
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.1
jupyter_core==5.1.5
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mtcnn==0.1.1
nest-asyncio==1.5.6
numpy==1.24.0
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.1
Pillow==9.3.0
pip==22.3.1
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
pyzmq==25.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
setuptools==65.5.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
tensorflow-metal==0.7.0
termcolor==2.1.1
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.11.0
